+++ app.js
@@ -4,7 +4,7 @@ const logger = require('morgan');
 const cors = require('cors');
 require('dotenv').config();
 
-const contactsRouter = require('./routes/api/contacts');
+const contactsRouter = require('./routes/api');
 
 const app = express();
