+++ controllers/contacts/removeById.js
@@ -1,0 +1,15 @@
+const {Contact} = require('@root/models');
+const {HttpError} = require('@root/helpers');
+
+const removeById = async (req, res) => {
+	const {id} = req.params;
+	const result = await Contact.findByIdAndRemove(id);
+	if (!result) {
+		throw HttpError(404, "Not found");
+	}
+	res.json({
+		message: "Contact deleted"
+	})
+};
+
+module.exports = removeById;
