+++ controllers/index.js
@@ -1,0 +1,21 @@
+const contacts = require('./contacts');
+
+module.exports = contacts;
+
+// const {
+// 	getAll,
+// 	getById,
+// 	add,
+// 	removeById,
+// 	updateById,
+// 	updateFavouriteById
+// } = require('./contacts');
+//
+// module.exports = {
+// 	getAll,
+// 	getById,
+// 	add,
+// 	removeById,
+// 	updateById,
+// 	updateFavouriteById
+// };
