+++ models/index.js
@@ -1,0 +1,9 @@
+const {
+	Contact,
+	schemas,
+} = require('./contact');
+
+module.exports = {
+	Contact,
+	schemas
+}
